Add basic setup.sh script
This script starts by checking if homebrew and oh-my-zsh are installed.

If either of these dependencies are missing it will exit early.

If all dependencies are found, then it will symlink a subset of the dotfiles
stored in this repo. I am using a subset because I am unsure whether I care
about any other dotfiles at this stage.